Replace jcenter with mavenCentral
Gradle 7 gives the following warning:

The RepositoryHandler.jcenter() method has been deprecated. This is scheduled to be removed in Gradle 8.0. JFrog announced JCenter's sunset in February 2021. Use mavenCentral() instead. Consult the upgrading guide for further information: https://docs.gradle.org/7.5/userguide/upgrading_version_6.html#jcenter_deprecation
        at build_25a2ym3sea7y143gert29dq0p$_run_closure1$_closure2.doCall(/home/derrick/.pub-cache/hosted/pub.dev/receive_intent-0.2.4/android/build.gradle:8)
        (Run with --stacktrace to get the full stack trace of this deprecation warning.)

The link states:
The jcenter() convenience method is now deprecated

JFrog announced the sunset of the JCenter repository in February 2021. Many Gradle builds rely on JCenter for project dependencies.

No new packages or versions are published to JCenter, but JFrog says they will keep JCenter running in a read-only state indefinitely. We recommend that you consider using mavenCentral(), google() or a private maven repository instead.

Gradle emits a deprecation warning when jcenter() is used as a repository and this method is scheduled to be removed in Gradle 8.0.